commit v2.0.0
* Translation Services are promoted in a different way within the WPML
Dashboard, therefore we have changed the plugin (and its name). This is
a temporary move, because WPML will come soon with a filter that
enables you to remove or hide the entire Translation Services block
from the Translators tab of the WPML Translation Management Addon.
* Changed the check for WPML being active into the Translation
Management addon being active instead.
* Changed the deactivation text
* Changed the plugin activation to remove any existing definitions from
the wp-config.php file